_site/scripts/requirements.txt
# MESGRO Scripts Dependencies
# ============================
# Install with: pip install -r requirements.txt
# Recommended: Use conda environment 'mesgro'
#   conda create -n mesgro python=3.11
#   conda activate mesgro
#   pip install -r requirements.txt

# Core dependencies (required)
trimesh>=4.0.0
numpy>=1.24.0

# STEP file support (required for .step/.stp files)
# Note: This is a large package (~100MB) with OpenCASCADE bindings
cadquery-ocp>=7.8.0

# Optional: Additional mesh format support
# pygltflib>=1.16.0  # For advanced GLTF manipulation
